+++ Runtime/Jobx/JobxConstants.cs
@@ -1,0 +1,11 @@
+using Unity.Mathematics;
+
+namespace Voxell.Jobx
+{
+  public static partial class Jobx
+  {
+    public const int S_BATCH_SIZE = 16;
+    public const int M_BATCH_SIZE = 64;
+    public const int L_BATCH_SIZE = 128;
+  }
+}