+++ Runtime/Jobx/SumScanJob.cs
@@ -10,23 +10,23 @@ namespace Voxell.Jobx
     [BurstCompile(CompileSynchronously = true)]
     private struct HillisSteeleSumScanJob : IJobParallelFor
     {
+      public int offset;
+
       [NativeDisableParallelForRestriction]
-      private NativeArray<int> na_array;
-
+      public NativeArray<int> na_array;
       [NativeDisableParallelForRestriction, ReadOnly]
-      private NativeArray<int> na_prevArray;
-      private int _offset;
+      public NativeArray<int> na_prevArray;
 
-      public HillisSteeleSumScanJob(ref NativeArray<int> na_array, ref NativeArray<int> na_prevArray, int offset)
+      public HillisSteeleSumScanJob(ref NativeArray<int> na_array, ref NativeArray<int> na_prevArray)
       {
         this.na_array = na_array;
         this.na_prevArray = na_prevArray;
-        _offset = offset;
+        this.offset = 1;
       }
 
       public void Execute(int index)
       {
-        int sumIdx = index - _offset;
+        int sumIdx = index - offset;
         if (sumIdx >= 0) na_array[index] += na_prevArray[sumIdx];
       }
     }
@@ -35,14 +35,19 @@ namespace Voxell.Jobx
     public static void InclusiveSumScan(NativeArray<int> na_array)
     {
       int arrayLength = na_array.Length;
+      JobHandle jobHandle;
+
       NativeArray<int> na_prevArray = new NativeArray<int>(na_array, Allocator.TempJob);
+      HillisSteeleSumScanJob sumScanJob = new HillisSteeleSumScanJob(ref na_array, ref na_prevArray);
+
       for (int offset=1; offset < arrayLength; offset <<= 1)
       {
-        HillisSteeleSumScanJob sumScanJob = new HillisSteeleSumScanJob(ref na_array, ref na_prevArray, offset);
-        JobHandle jobHandle = sumScanJob.Schedule(arrayLength, 16);
+        sumScanJob.offset = offset;
+        jobHandle = sumScanJob.Schedule(arrayLength, 64);
         jobHandle.Complete();
         na_prevArray.CopyFrom(na_array);
       }
+
       na_prevArray.Dispose();
     }
   }